Tie visual board to engine, prevent invalid plays
Add unique id's (markx) to each spot on visual board, and use that to
tie the visual board to the internal board in game logic. Add more
checks to validMove function to fully prevent invalid moves on already
played spot or after game is over. All notifications of player turn,
and win/draw status are still in console.