Create game logic for subGame of MegaTTT
Create megaEngine.js to house game logic for new game mode MEGA tic tac
toe. Create masterGame object to house the meta-game, as well as an
array of the nine sub-games. Copy over several functions from the basic
game engine and rewrite portions of them to fit the manipulation of
a nested subGame. suGames are now playable in the console.